move rules under rules package
