+++ service/controller/resource/alert/create.go
@@ -8,13 +8,13 @@ import (
 )
 
 func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
-	serviceMonitors, err := toServiceMonitors(obj)
+	prometheusRules, err := toPrometheusRules(obj)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
-	for _, serviceMonitor := range serviceMonitors {
-		_, err = r.prometheusClient.MonitoringV1().ServiceMonitors(serviceMonitor.GetNamespace()).Create(serviceMonitor)
+	for _, prometheusRule := range prometheusRules {
+		_, err = r.prometheusClient.MonitoringV1().PrometheusRules(prometheusRule.GetNamespace()).Create(prometheusRule)
 		if apierrors.IsAlreadyExists(err) {
 			// fall through
 		} else if err != nil {
