+++ service/controller/resource/alert/resource.go
@@ -1,9 +1,11 @@
 package alert
 
 import (
+	promv1 "github.com/coreos/prometheus-operator/pkg/apis/monitoring/v1"
 	promclient "github.com/coreos/prometheus-operator/pkg/client/versioned"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
+	"github.com/giantswarm/prometheus-meta-operator/service/key"
 )
 
 const (
@@ -38,4 +40,17 @@ func New(config Config) (*Resource, error) {
 
 func (r *Resource) Name() string {
 	return Name
+}
+
+func toPrometheusRules(obj interface{}) ([]*promv1.PrometheusRule, error) {
+	cluster, err := key.ToCluster(obj)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	rules := []*promv1.PrometheusRule{
+		rules.APIServer(),
+	}
+
+	return rules, nil
 }
